673 implement a heuristic to detect whether the ruff formatter is used (#675)
* Add a heuristic to determine whether the Ruff linter is being used.

This is provided as a new method on the RuffTool class called `is_linter_used`.

The heuristic relies on the presence of a `lint` section in the Ruff configuration.

Some refactoring was required to achieve this, including splitting `is_config_present` in the Tool class into a new helper method `_is_config_spec_present`

Also there seemed to have been a bug where the `is_config_present` method would try and access files that don't exist instead of skipping them for IO

* Add a `is_formatter_used` method to `RuffTool` similar to `is_linter_used`.