Remove idempotent encode-decode calls for unicode character
I think this was added from a contributor without testing it was necessary. It seems to have no effect, so I've removed it. In hindsight, it was fairly clear it had no effect because the other unicode characters didn't use it.

I also noticed an issue regarding missing messages when looking into this
https://github.com/usethis-python/usethis-python/issues/881